Refactored AuthButton to be more readable - Made dropdown a separate componenet
