+++ src/components/Auth/AuthButton.js
@@ -6,12 +6,11 @@ import { withRouter, Link } from "react-router-dom"
 import { useAuth0 } from "./react-auth0-spa.js"
 import MenuItem from "@material-ui/core/MenuItem"
 import { Popover } from "@material-ui/core"
+import Dropdown from "./Dropdown"
 
-const AuthButton = ({ first_name, pictureURL, history }) => {
-  const { loginWithRedirect, logout } = useAuth0()
+const AuthButton = () => {
+  const { loginWithRedirect } = useAuth0()
   const [anchorEl, setAnchorEl] = React.useState(null)
-
-  const { push } = history
 
   const handleClick = e => {
     setAnchorEl(e.currentTarget)
@@ -21,19 +20,6 @@ const AuthButton = ({ first_name, pictureURL, history }) => {
     setAnchorEl(null)
   }
 
-  const pushToRoute = () => {
-    console.log("trigger")
-    console.log("history", history)
-    push("/dashboard")
-  }
-
-  const logoutWithRedirect = () => {
-    Cookie.remove("user_id")
-    logout({
-      returnTo: window.location.origin,
-    })
-  }
-
   const userIDCookie = Cookie.get("user_id")
 
   return (
@@ -41,74 +27,19 @@ const AuthButton = ({ first_name, pictureURL, history }) => {
       {
         //displays sign in button if no user cookie found
       }
-      {!userIDCookie && (
-        <>
-          <Button onClick={() => loginWithRedirect({})} color="secondary">
-            Sign In
-          </Button>
-        </>
-      )}
-      {
-        //displays sign out button if user cookie found
-      }
-      {userIDCookie && (
-        <>
-          <Button
-            aria-controls="simple-menu"
-            aria-haspopup="true"
-            onClick={handleClick}
-          >
-            {/* if there is not a pictureURL, the material UI account icon appears */}
-            {!pictureURL ? (
-              <i id="account-box" class="material-icons">
-                account_box
-              </i>
-            ) : (
-                <img
-                  className="user-image"
-                  src={pictureURL}
-                  alt="user profile"
-                ></img>
-              )}
-
-            <i id="account-arrow" className="material-icons">
-              {anchorEl === null ? "keyboard_arrow_down" : "keyboard_arrow_up"}
-            </i>
-          </Button>
-          <Popover
-            className="list-example"
-            id="simple-menu"
-            anchorEl={anchorEl}
-            keepMounted
-            open={Boolean(anchorEl)}
-            onClose={handleClose}
-            anchorOrigin={{ vertical: "bottom", horizontal: "left" }}
-            transformOrigin={{ vertical: "top", horizontal: "left" }}
-          >
-            <MenuItem onClick={() => history.push('/dashboard/profile')}>My Profile</MenuItem>
-            <MenuItem onClick={() => history.push('/dashboard')}>Dashboard</MenuItem>
-            <MenuItem
-              color="secondary"
-              onClick={() => {
-                logoutWithRedirect()
-              }}
-            >
-              Sign Out
-            </MenuItem>
-          </Popover>
-        </>
+      {!userIDCookie ? (
+        <Button onClick={() => loginWithRedirect({})} color="secondary">
+          Sign In
+        </Button>
+      ) : (
+        <Dropdown
+          anchorEl={anchorEl}
+          handleClose={handleClose}
+          handleClick={handleClick}
+        />
       )}
     </>
   )
 }
 
-const mapStateToProps = ({ usersReducer }) => {
-  return { ...usersReducer.user }
-}
-
-export default withRouter(
-  connect(
-    mapStateToProps,
-    {}
-  )(AuthButton)
-)
+export default AuthButton
